requirements.txt
asgiref==3.6.0
attrs==23.1.0
build==0.10.0
CacheControl==0.12.11
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
colorama==0.4.6
crashtest==0.4.1
cryptography==41.0.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
dulwich==0.21.5
envparse==0.2.0
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.10.1
filelock==3.12.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
more-itertools==9.1.0
msgpack==1.0.5
oauthlib==3.2.2
packaging==23.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.0.0
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.1
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
pywin32-ctypes==0.2.0
rapidfuzz==2.15.1
requests==2.30.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
shellingham==1.5.0.post1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.5.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.1
webencodings==0.5.1
zipp==3.15.0
